--- cpp_effects_from_scratch/handleAudio.h
@@ -1,8 +1,0 @@
-#include "wavHeader.cpp"
-#include <vector>
-#include <iostream>
-#include <fstream>
-
-std::vector<float> convertBytesToSamples16(const std::vector<uint8_t>& audioData);
-std::vector<uint8_t> convertSamplesToBytes16(const std::vector<float>& samples);
-void saveWavFile(const char* filename, const WavHeader& header, const std::vector<uint8_t>& audioBytes);
