+++ cpp_effects_from_scratch/include/audio/wavHeader.h
@@ -1,0 +1,22 @@
+#ifndef WAVHEADER_H
+#define WAVHEADER_H
+
+#include <cstdint>
+
+struct WavHeader {
+    char riff[4];
+    uint32_t chunkSize;
+    char wave[4];
+    char fmt[4];
+    uint32_t subchunk1Size;
+    uint16_t audioFormat;
+    uint16_t numChannels;
+    uint32_t sampleRate;
+    uint32_t byteRate;
+    uint16_t blockAlign;
+    uint16_t bitsPerSample;
+    char data[4];
+    uint32_t dataSize;
+};
+
+#endif // WAVHEADER_H
